### 2022-06-26 #4509 @editorialbot
[PRE REVIEW]: PyFPT: A Python package for first-passage times
**Submitting author:** @Jacks0nJ (<a href="http://orcid.org/0000-0002-1564-1607">Joseph Jackson</a>)
**Repository:** https://github.com/Jacks0nJ/PyFPT
**Branch with paper.md** (empty if default branch):
**Version:** 1.00
