+++ paper/paper.bib
@@ -104,8 +104,12 @@ Doi = {10.1140/epjb/s10051-021-00246-0},
     archivePrefix = "arXiv",
     primaryClass = "astro-ph.CO",
     reportNumber = "HIP-2021-31/TH",
-    month = "11",
-    year = "2021"
+    doi = "10.1088/1475-7516/2022/05/027",
+    journal = "JCAP",
+    volume = "05",
+    number = "05",
+    pages = "027",
+    year = "2022"
 }
 
 
