### 2022-07-01 #4509 @drvinceknight
@CFGrote @FaroukM @OsAmaro I'm pinging you to ask if you would be able to assist with reviewing this submissions to JOSS (The Journal of Open Source Software).

JOSS publishes articles about open source research software. The submission I'd like you to review is titled: “PyFPT: A Python package for first-passage times”. You can find more information at the top of this Github issue.
