### 2022-07-18 Jacks0nJ/PyFPT#34 @CFGrote
[JOSS Review]: Usage example
Re  [JOSS #4509](https://github.com/openjournals/joss-reviews/issues/4509)

## Tutorial notebook

The documentation at  https://pyfpt.readthedocs.io/en/latest/Installing%20and%20First%20Simulation.html sources the User guide notebooks in
`User guides/`. I found no issue going through them.

 A minor source of confusion may arise from the additional (redundant?) notebooks in `docs/source`, which look identical but miss some
local dependencies (e.g. the mpl_style.py module). Maybe they can be removed?
