### 2022-07-18 Jacks0nJ/PyFPT#35 @CFGrote
[JOSS Review] User Guide 2: Understanding the Data
Re [JOSS #4509](https://github.com/openjournals/joss-reviews/issues/4509)

Running the second user guide notebook, cell 9 aborts with

```
for i, bias_amp in enumerate(bias_amp_range):
    print('Now simulating for bias amplitude '+str(bias_amp))
    bin_centres, heights, errors =\
        fpt.numerics.is_simulation(drift_func, diffusion_func, phi_in, phi_end,
                                   num_runs, bias_amp, dN, save_data=True,
                                   estimator='lognormal', bins=num_bins)

    # Storing the results in NumPy arrays
    bin_centres_storage[:len(bin_centres),i] = bin_centres
    heights_storage[:len(heights),i] = heights
    # Remeber this is a 3D array
    errors_storage[i,:,:len(heights)] = errors
```
```
/opt/anaconda3/envs/fpt_review/lib/python3.10/site-packages/numpy/lib/function_base.py:2830: RuntimeWarning: invalid value encountered in true_divide
  c /= stddev[None, :]

Saved data to file IS_data_x_in_6.48_iterations_200000_bias_0.0.csv

Input In [9], in <cell line: 1>()
      1 for i, bias_amp in enumerate(bias_amp_range):
      2     print('Now simulating for bias amplitude '+str(bias_amp))
      3     bin_centres, heights, errors =\
----> 4         fpt.numerics.is_simulation(drift_func, diffusion_func, phi_in, phi_end,
      5                                    num_runs, bias_amp, dN, save_data=True,
      6                                    estimator='lognormal', bins=num_bins)
      8     # Storing the results in NumPy arrays
      9     bin_centres_storage[:len(bin_centres),i] = bin_centres

File /opt/anaconda3/envs/fpt_review/lib/python3.10/site-packages/pyfpt/numerics/is_simulation.py:229, in is_simulation(drift, diffusion, x_in, x_end, num_runs, bias, time_step, bins, min_bin_size, num_sub_samples, estimator, save_data, t_in, t_f, x_r)
    224         save_data_to_file(fpt_values, w_values, x_in, num_runs,
    225                           bias(x_in, 0), extra_label='_custom_bias')
    227 # Now analysisng the data to creating the histogram/PDF data
    228 bin_centres, heights, errors, num_runs_used, bin_edges_untruncated =\
--> 229     data_points_pdf(fpt_values, w_values, estimator, bins=bins,
    230                     min_bin_size=min_bin_size,
    231                     num_sub_samples=num_sub_samples)
    232 # Return data as lists
    233 return bin_centres.tolist(), heights.tolist(), errors.tolist()

File /opt/anaconda3/envs/fpt_review/lib/python3.10/site-packages/pyfpt/numerics/data_points_pdf.py:121, in data_points_pdf(data, weights, estimator, bins, min_bin_size, num_sub_samples)
    119 if estimator == 'naive':
    120     heights = heights_raw/histogram_norm
--> 121     errors = jackknife_errors(data, weights, bins, num_sub_samples)
    122     if isinstance(min_bin_size, int) is True:
    123         heights = heights[filled_bins]

File /opt/anaconda3/envs/fpt_review/lib/python3.10/site-packages/pyfpt/numerics/jackknife_errors.py:53, in jackknife_errors(data_input, weights_input, bins, num_sub_samps)
     49 height_array = np.zeros((num_bins, num_sub_samps))  # Storage
     51 # Next organise into subsamples
     52 data =\
---> 53     np.reshape(data, (int(data.shape[0]/num_sub_samps), num_sub_samps))
     54 weights =\
     55     np.reshape(weights, (int(weights.shape[0]/num_sub_samps),
     56                          num_sub_samps))
     58 # Find the heights of the histograms, for each sample

File <__array_function__ internals>:180, in reshape(*args, **kwargs)

File /opt/anaconda3/envs/fpt_review/lib/python3.10/site-packages/numpy/core/fromnumeric.py:298, in reshape(a, newshape, order)
    198 @array_function_dispatch(_reshape_dispatcher)
    199 def reshape(a, newshape, order='C'):
    200     """
    201     Gives a new shape to an array without changing its data.
    202
   (...)
    296            [5, 6]])
    297     """
--> 298     return _wrapfunc(a, 'reshape', newshape, order=order)

File /opt/anaconda3/envs/fpt_review/lib/python3.10/site-packages/numpy/core/fromnumeric.py:57, in _wrapfunc(obj, method, *args, **kwds)
     54     return _wrapit(obj, method, *args, **kwds)
     56 try:
---> 57     return bound(*args, **kwds)
     58 except TypeError:
     59     # A TypeError occurs if the object does have such a method in its
     60     # class, but its signature is not identical to that of NumPy's. This
   (...)
     64     # Call _wrapit from within the except clause to ensure a potential
     65     # exception has a traceback chain.
     66     return _wrapit(obj, method, *args, **kwds)

ValueError: cannot reshape array of size 199968 into shape (9998,20)
```

Small typo in the MD cell after input [5]: "Now do let's do a 2~~H~~D histogram as a colour map."
