### 2022-07-18 Jacks0nJ/PyFPT#34 @Jacks0nJ
I agree this could lead to confusion. The reason they are duplicated is the user guides in docs/source are used to generate html pages for the documentation, so the user guides can also be viewed again there. It was easier to simply duplicate them where the ReadTheDocs pages are generated. I'm more than happy to implement a more elegant solution if you have suggestion.
