### 2022-07-18 Jacks0nJ/PyFPT#34 @Jacks0nJ
@CFGrote I'll look into it. I'm currently away from my linux machine, but will try to implement it once I'm back.
