### 2022-07-18 Jacks0nJ/PyFPT#36 @CFGrote
[JOSS Review] Testing
Re [JOSS #4509](https://github.com/openjournals/joss-reviews/issues/4509)

## Testing
I could not find instructions for how to run tests. I found that running
```console
$> pytest -v
```
in the root dir works, though.

`tests/test_is_simulation.py::TestIS_Simulation::test_is_simulation`
generates two matplolib plot windows that need to be closed manually before the tests continue. I find that slightly annoying as tests cannot be run automatically. As I'd like to recommended adding the tests to a CI setup (e.g. github-actions), this would have to be addressed. One solution could be to check if the test is run interactively or within a CI automated setup and have the test generate these plots or not.

I  suggest adding information about how to run tests to the README and documentation.

Also, there seems to be some coverage testing functionality as indicated by lines 16,17 in `setup.cfg`, would be great to learn how to run coverage testing and get a coverage report. I assume coverage is great (looking at the impressive number of test modules), but I'd like to see the actual numbers ;)
