### 2022-07-18 #4509 @CFGrote
This concludes my first round of review. I'm really impressed by the quality of the code and documentation and for following OS Dev best practices in general. I'm looking forward to your response @Jacks0nJ .
