### 2022-07-18 #4509 @Jacks0nJ
Cheers @CFGrote! I'll try to work on your feedback over the next two days but I'm on holiday 20/07/22-29/07/22, so just to let you know some fixes might come after then.

If there is anything else I should do @drvinceknight apart from what @CFGrote has already highlighted, do let me know.
