### 2022-07-18 675d7cb
Update paper.md
Split paragraph
