### 2022-07-26 #4607 @editorialbot
[REVIEW]: PyFPT: A Python package for first-passage times
**Submitting author:** @Jacks0nJ (<a href="http://orcid.org/0000-0002-1564-1607">Joseph Jackson</a>)
**Repository:** https://github.com/Jacks0nJ/PyFPT
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.2.1
