### 2022-08-02 f07b9f5
Implemented all of JOSS bar coverage
