+++ requirements.txt
@@ -6,3 +6,6 @@ mpmath
 matplotlib
 scipy
 nbsphinx
+nbsphinx-link
+coverage
+pytest-cov
