+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = PyFPT
-version = 1.0.0
+version = 1.0.1s
 author = Joe Jackson
 author_email = joseph.jackson@port.ac.uk
 description = Stochastic first-passage time (FPT) simulations using importance sampling
