### 2022-08-02 1e5b37e
Removed s from .cfg
