### 2022-08-02 9db430f
fixed / to % to find remiander
