### 2022-08-02 Jacks0nJ/PyFPT#34 @Jacks0nJ
I've used the [nbsphinx-link](https://github.com/vidartf/nbsphinx-link) to allow the documentation read directly from the guides folder. It seems wanting to add Jupyter notebooks to the documentation from a guides folder is a common problem!

So the duplication should hopefully be fixed.
