### 2022-08-02 Jacks0nJ/PyFPT#35 @Jacks0nJ
I've struggled to find the source of the error you have experienced here, as I have been unable to reproduce it. There is usually a different but minor error which occurs here as the code is really designed for a non-zero bias, which I'll fix in a future patch, but the error you found is strange. It both `should' not have occurred for the simulation done (it can occur in diffusion domination) and I had put in some safe guards against it...

Anyway, I've now added additional safe guards so hopefully it won't happen again. Could you rerun it again to test please? Oh and make sure you are using PyFPT 1.0.1.

Do let me know if it is still occurring.

I've also fixed the typo :)
