### 2022-08-03 cf9d33b
Delete update-coverage-on-readme.yml
