### 2022-08-03 Jacks0nJ/PyFPT#36 @Jacks0nJ
I've now added instructions for testing the code in the README.

The two plots being being generated during testing has now been suppressed. The code already has CI setup; the plots appearing are a rare event to alert the user there could be a problem, so simply by chance had not been generated during previous testing of the code using unit tests.

As for coverage testing, I've struggled to implement it. Do you have a recommended method?

(The problem has been I either need to use a third-party provided, which I don't really want to do, or the locally done coverage tests simply break. It seems to be to do with how I set up testing; locally run tests work on the installed package (rather than local code), such that it tests the uploaded code when I push to Github (this is becuase it needs to compile Cython). But it appears coverage.py (and similair packages) can't find the functions to be tested, as it is not in the directory, e.g. test_is_simulation can't find pyfpt.numerics.is_simulation when I run coverage tests.)
