### 2022-08-18 #4607 @Jacks0nJ
@drvinceknight
I've been updating the code to also have functionality for n-dimensions. Currently, the main branch is only for 1D. Apart from a few functions being renamed for clarity and some efficiency savings I've found, the old 1D code is untouched and relevant guides will be the same. There would also be a new, 4th guide to explain the nD code.

The nD code is almost ready to merge with the main branch (along with the updated guides). Should I hold off on doing this until after the review is completed? Or would it be possible to have the JOSS paper be for the nD code?
