### 2022-08-23 Jacks0nJ/PyFPT#36 @CFGrote
Ok, I'm back :sun_with_face::tropical_drink:

What works for me is:
```
coverage run -m pytest tests/
```
to run the coverage test, followed by
```
coverage report
```
to generate the report:
```
Name                                                    Stmts   Miss  Cover
pyfpt/__init__.py                                           2      0   100%
pyfpt/analytics/__init__.py                                18      0   100%
pyfpt/analytics/classicality_criterion.py                  12      0   100%
pyfpt/analytics/edgeworth_pdf.py                           28      0   100%
pyfpt/analytics/fourth_central_moment_efolds.py            19      0   100%
pyfpt/analytics/gaussian_deviation.py                      31      1    97%
pyfpt/analytics/gaussian_pdf.py                            12      0   100%
pyfpt/analytics/kurtosis_efolds.py                         10      1    90%
pyfpt/analytics/mean_efolds.py                             20      0   100%
pyfpt/analytics/optimal_bias_amplitude.py                  38     17    55%
pyfpt/analytics/quadratic_inflation_large_mass_pdf.py      45     37    18%
pyfpt/analytics/quadratic_inflation_near_tail_pdf.py       28     23    18%
pyfpt/analytics/reduced_potential.py                        8      0   100%
pyfpt/analytics/reduced_potential_ddiff.py                  8      0   100%
pyfpt/analytics/reduced_potential_diff.py                   8      0   100%
pyfpt/analytics/skewness_efolds.py                         22      0   100%
pyfpt/analytics/slow_roll_diffusion.py                      6      0   100%
pyfpt/analytics/slow_roll_drift.py                          4      0   100%
pyfpt/analytics/third_central_moment_efolds.py             20      0   100%
pyfpt/analytics/variance_efolds.py                         20      0   100%
pyfpt/numerics/__init__.py                                 12      0   100%
pyfpt/numerics/data_in_histogram_bins.py                   19      0   100%
pyfpt/numerics/data_points_pdf.py                          54      4    93%
pyfpt/numerics/histogram_data_truncation.py                31      1    97%
pyfpt/numerics/histogram_normalisation.py                   6      1    83%
pyfpt/numerics/is_simulation.py                            82     32    61%
pyfpt/numerics/jackknife_errors.py                         31      4    87%
pyfpt/numerics/log_normal_error.py                         10      0   100%
pyfpt/numerics/log_normal_height.py                         4      0   100%
pyfpt/numerics/log_normal_mean.py                           7      0   100%
pyfpt/numerics/lognormality_check.py                       42     18    57%
pyfpt/numerics/multi_processing_error.py                    8      0   100%
pyfpt/numerics/save_data_to_file.py                        12      1    92%
tests/test_classicality_criterion.py                       22      1    95%
tests/test_data_in_histogram_bins.py                       16      1    94%
tests/test_data_points_pdf.py                              28      1    96%
tests/test_edgeworth_pdf.py                                27      1    96%
tests/test_fourth_central_moment_efolds.py                 23      1    96%
tests/test_gaussian_deviation.py                           23      1    96%
tests/test_gaussian_pdf.py                                 27      1    96%
tests/test_histogram_data_truncation.py                    25      1    96%
tests/test_histogram_normalisation.py                      18      1    94%
tests/test_importance_sampling_cython.py                   56      1    98%
tests/test_is_simulation.py                                39      1    97%
tests/test_jackknife_errors.py                             24      1    96%
tests/test_kurtosis_efolds.py                              23      1    96%
tests/test_log_normal_error.py                             14      1    93%
tests/test_log_normal_height.py                            14      1    93%
tests/test_log_normal_mean.py                              13      1    92%
tests/test_lognormality_check.py                           35      1    97%
tests/test_mean_efolds.py                                  23      1    96%
tests/test_multi_processing_error.py                       17      1    94%
tests/test_optimal_bias_amplitude.py                       24      1    96%
tests/test_reduced_potential.py                            14      1    93%
tests/test_reduced_potential_ddiff.py                      14      1    93%
tests/test_reduced_potential_diff.py                       14      1    93%
tests/test_save_data_to_file.py                            29      1    97%
tests/test_skewness_efolds.py                              23      1    96%
tests/test_slow_roll_diffusion.py                          28      5    82%
tests/test_slow_roll_drift.py                              26      3    88%
tests/test_third_central_moment_efolds.py                  23      1    96%
tests/test_variance_efolds.py                              23      1    96%
TOTAL                                                    1362    175    87%
```

Not bad, actually :smiley:

But maybe I don't understand your exact problem?
