### 2022-08-23 Jacks0nJ/PyFPT#36 @Jacks0nJ
Great, hope you had a good holiday!

And I'm glad the unit tests are working!

Sorry, I thought you were originally suggesting having a coverage badge on the README which automatically displays the coverage. But this was proving very difficult and your method is a lot easier.
