### 2022-08-23 Jacks0nJ/PyFPT#36 @CFGrote
> Great, hope you had a good holiday!
Totally, although a bit short, as always.
>
> And I'm glad the unit tests are working!
Yep
>
> Sorry, I thought you were originally suggesting having a coverage badge on the README which automatically displays the coverage. But this was proving very difficult and your method is a lot easier.
No, that was not my intention. I think we can close this one now.
