### 2022-09-27 Jacks0nJ/PyFPT#35 @CFGrote
Small typo in MD headed "Interpreting the Data and Lognormal Method (estimator)": ~normalsied~ normalised
