### 2022-09-27 Jacks0nJ/PyFPT#35 @CFGrote
Otherwise all good, the notebook completes without errors.
