### 2022-10-11 #4607 @CFGrote
IMO you should wait with merging until the JOSS paper is out. This is a major change and warrants a major version increment. That's my point of view, but I'm happy to discuss.
