### 2022-10-24 #4607 @Jacks0nJ
Any ETA on when the review will be finished? As I've already had interest in the 2D version of the code, which I would like to be available in the main branch and through PyPI (rather than TestPyPI as is currently the case).  Additionally I will be presenting a tutorial on the code next month, so ideally I would have the merged the multi-dimensional version of the code before then.
