### 2022-10-31 0ebac91
changed returned values to lists
