+++ tests/test_re_processing.py
@@ -2,6 +2,7 @@ import unittest
 import numpy as np
 
 from pyfpt.numerics import re_processing
+from scipy.interpolate import interp1d
 
 
 class TestReProcessing(unittest.TestCase):
@@ -26,19 +27,18 @@ class TestReProcessing(unittest.TestCase):
 
         bin_centres_naive = np.array(bin_centres_naive)
         heights_naive = np.array(heights_naive)
-        # Only check filled bins, which can be determined by comparing the bin
-        # centre values. Checking if the truncation is correct is another test
-        # script.
-        diff_naive = np.zeros(len(heights_naive))
-        for i, bin_centre in enumerate(bin_centres_naive):
-            # The smallest differance should correspond to the same bin
-            bin_diff = np.abs((bin_centres_expected-bin_centre)/bin_centre)
-            j = np.argmin(bin_diff)
-            diff_naive[i] =\
-                np.abs((heights_naive[i]-heights_expected[j]) /
-                       heights_expected[j])
-        # Now check that differance is very small, as it should be identical
-        self.assertTrue(all(diff_naive < 0.00001))
+        # As the bin centres might be different, need to use an interpolation
+        # to compare
+        interpolatent = interp1d(bin_centres_expected, heights_expected)
+        # Need to have the unfilled values not be 0
+        diff_naive = np.zeros(len(heights_naive))+1
+        for i in range(len(heights_naive)):
+            expected = float(interpolatent(bin_centres_naive[i]))
+            if expected < 10**-5:
+                diff_naive[i] = np.abs((expected-heights_naive[i]))
+            else:
+                diff_naive[i] = np.abs((expected-heights_naive[i])/expected)
+        self.assertTrue(all(diff_naive < 0.1))
 
         # Testing this function y using mock data, where although data is
         # weighted, the outcome should be the same as if all weights were 1.
@@ -70,20 +70,20 @@ class TestReProcessing(unittest.TestCase):
                           bins=num_bins)
         bin_centres_naive = np.array(bin_centres_naive)
         heights_naive = np.array(heights_naive)
-        # Only check filled bins, which can be determined by comparing the bin
-        # centre values. Checking if the truncation is correct is another test
-        # script.
-        diff_naive = np.zeros(len(heights_naive))
-        for i, bin_centre in enumerate(bin_centres_naive):
-            # The smallest differance should correspond to the same bin
-            bin_diff = np.abs((bin_centres_expected-bin_centre)/bin_centre)
-            j = np.argmin(bin_diff)
-            diff_naive[i] =\
-                np.abs((heights_naive[i]-heights_expected[j]) /
-                       heights_expected[j])
+        # As the bin centres might be different, need to use an interpolation
+        # to compare
+        interpolatent = interp1d(bin_centres_expected, heights_expected)
+        # Need to have the unfilled values not be 0
+        diff_naive = np.zeros(len(heights_naive))+1
+        for i in range(len(heights_naive)):
+            expected = float(interpolatent(bin_centres_naive[i]))
+            if expected < 10**-5:
+                diff_naive[i] = np.abs((expected-heights_naive[i]))
+            else:
+                diff_naive[i] = np.abs((expected-heights_naive[i])/expected)
         # Now check that differance is small. A small differance is expected
         # as comparing weighted data against unweighted expectation.
-        self.assertTrue(all(diff_naive < 0.05))
+        self.assertTrue(all(diff_naive < 0.1))
 
         # Now let's do the same for the lognormal distribution
 
@@ -93,19 +93,16 @@ class TestReProcessing(unittest.TestCase):
                           bins=num_bins)
         bin_centres_lognormal = np.array(bin_centres_lognormal)
[…]
