### 2022-11-22 #4607 @drvinceknight
Everything looks good to me @Jacks0nJ. The suggestion made by editorial bot does seem correct to me, would you be able to add the doi `10.1088/1475-7516/2022/09/045` which points at https://iopscience.iop.org/article/10.1088/1475-7516/2022/09/045 (unless I'm wrong and it differs from the arXiv preprint).

Let me know once that's done and there will be a couple of other steps to take but it won't take us long.
