+++ paper/paper.bib
@@ -93,7 +93,10 @@ Doi = {10.1140/epjb/s10051-021-00246-0},
     eprint = "2204.03859",
     archivePrefix = "arXiv",
     primaryClass = "astro-ph.CO",
-    month = "4",
+    doi = "10.1088/1475-7516/2022/09/045",
+    journal = "JCAP",
+    volume = "09",
+    pages = "045",
     year = "2022"
 }
 
