### 2022-11-23 05528dd
minor changes
swapped two sentences around and added  "in general" for clarity.
