### 2022-11-23 #4607 @Jacks0nJ
@drvinceknight I've added the doi to the reference. The pre-print version does in fact differ from the published version in content,  but the key point of their code not being open source is till valid.

I've also made some very minor changes: I've changed my name to "Joseph H. P. Jackson", swapped the second and third sentence around, added "in general" and added an additional grant number to the acknowledgments. Feel free to check the changes are ok and are indeed minor.

Do let me know if there is anything else I need to do.
