### 2022-11-24 #4607 @Jacks0nJ
Mind linking to a guide on how to do this? Apologies but I'm slightly unsure what tagged release is.

Do I just need to follow the steps outlined in the link below?

 https://docs.github.com/en/repositories/releasing-projects-on-github/managing-releases-in-a-repository
