### 2022-11-25 #4607 @drvinceknight
No problem @Jacks0nJ (no need to apologise), the steps here are indeed what you need to do: https://docs.github.com/en/repositories/releasing-projects-on-github/managing-releases-in-a-repository#creating-a-release
