### 2022-12-09 #4607 @drvinceknight
You can make an archive using Zenodo: https://zenodo.org

You can do this manually (although a lot of people set up something automatic between their repos). Once you create the archive (you'll need to upload a version of the release) you'll have a DOI which you can just put here. Let me know.
