### 2022-12-14 #4607 @drvinceknight
Sorry! But I'm afraid the archive needs to be modified a bit: the title needs to be `PyFPT: A Python package for first-passage times` the same as the article. I believe you can edit the archive directly on Zenodo without the need for a new release.
