### 2023-01-13 #4607 @danielskatz
@Jacks0nJ - I'm the track editor, handling the final processing.  I've suggested some small changes in https://github.com/Jacks0nJ/PyFPT/pull/47 - please merge this, or let me know what you disagree with, then we can proceed.
